requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-fasthtml==0.12.1
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
passlib==1.7.4
qrcode==8.2
SQLAlchemy==2.0.41
asyncpg==0.30.0
